Remove obsolete controller tests (#349)
