--- tests/Functional/Controller/Source/GitSourceControllerTest.php
@@ -1,80 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Controller\Source;
-
-use App\Tests\Application\AbstractApplicationTestCase;
-use App\Tests\DataProvider\ServiceBadResponseContentTypeDataProviderTrait;
-use App\Tests\DataProvider\ServiceHttpFailureDataProviderTrait;
-use App\Tests\Functional\Controller\AssertJsonResponseTrait;
-use App\Tests\Functional\GetClientAdapterTrait;
-use GuzzleHttp\Handler\MockHandler;
-use Psr\Http\Client\ClientInterface as HttpClientInterface;
-use Psr\Http\Message\ResponseInterface;
-use SmartAssert\UsersClient\ClientInterface as UsersClient;
-use SmartAssert\UsersClient\Model\Token;
-
-class GitSourceControllerTest extends AbstractApplicationTestCase
-{
-    use GetClientAdapterTrait;
-    use AssertJsonResponseTrait;
-    use ServiceBadResponseContentTypeDataProviderTrait;
-    use ServiceHttpFailureDataProviderTrait;
-
-    /**
-     * @dataProvider serviceExceptionDataProvider
-     *
-     * @param array<mixed> $expectedData
-     */
-    public function testActHandlesException(
-        \Exception|ResponseInterface $httpFixture,
-        int $expectedStatusCode,
-        array $expectedData
-    ): void {
-        $mockingHttpClient = self::getContainer()->get('app.test.mocking_http_client');
-        \assert($mockingHttpClient instanceof HttpClientInterface);
-
-        $httpMockHandler = self::getContainer()->get(MockHandler::class);
-        \assert($httpMockHandler instanceof MockHandler);
-
-        $httpMockHandler->append($httpFixture);
-
-        $apiKey = md5((string) rand());
-        $apiToken = md5((string) rand());
-        $label = md5((string) rand());
-        $hostUrl = md5((string) rand());
-        $path = md5((string) rand());
-        $credentials = md5((string) rand());
-
-        $usersClient = \Mockery::mock(UsersClient::class);
-        $usersClient
-            ->shouldReceive('createApiToken')
-            ->with($apiKey)
-            ->andReturn(new Token($apiToken))
-        ;
-
-        self::getContainer()->set(UsersClient::class, $usersClient);
-        self::getContainer()->set(HttpClientInterface::class, $mockingHttpClient);
-
-        $response = $this->applicationClient->makeCreateGitSourceRequest(
-            $apiKey,
-            $label,
-            $hostUrl,
-            $path,
-            $credentials
-        );
-        $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function serviceExceptionDataProvider(): array
-    {
-        return array_merge(
-            $this->serviceBadResponseContentTypeDataProvider('source', 'application/json'),
-            $this->serviceHttpFailureDataProvider('source'),
-        );
-    }
-}
