--- tests/Functional/Controller/User/RefreshTokenControllerTest.php
@@ -1,89 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Controller\User;
-
-use App\Tests\Application\AbstractApplicationTestCase;
-use App\Tests\DataProvider\ServiceBadResponseContentTypeDataProviderTrait;
-use App\Tests\DataProvider\ServiceHttpFailureDataProviderTrait;
-use App\Tests\Functional\Controller\AssertJsonResponseTrait;
-use App\Tests\Functional\GetClientAdapterTrait;
-use GuzzleHttp\Handler\MockHandler;
-use Psr\Http\Client\ClientInterface as HttpClientInterface;
-use Psr\Http\Message\ResponseInterface;
-
-class RefreshTokenControllerTest extends AbstractApplicationTestCase
-{
-    use GetClientAdapterTrait;
-    use AssertJsonResponseTrait;
-    use ServiceBadResponseContentTypeDataProviderTrait;
-    use ServiceHttpFailureDataProviderTrait;
-
-    /**
-     * @dataProvider serviceExceptionDataProvider
-     *
-     * @param array<mixed> $expectedData
-     */
-    public function testRevokeAllForUserHandlesException(
-        \Exception|ResponseInterface $httpFixture,
-        int $expectedStatusCode,
-        array $expectedData
-    ): void {
-        $mockingHttpClient = self::getContainer()->get('app.test.mocking_http_client');
-        \assert($mockingHttpClient instanceof HttpClientInterface);
-
-        $httpMockHandler = self::getContainer()->get(MockHandler::class);
-        \assert($httpMockHandler instanceof MockHandler);
-
-        $httpMockHandler->append($httpFixture);
-
-        $token = md5((string) rand());
-        $id = md5((string) rand());
-
-        self::getContainer()->set(HttpClientInterface::class, $mockingHttpClient);
-
-        $response = $this->applicationClient->makeRevokeAllRefreshTokensForUserRequest($token, $id);
-
-        $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
-    }
-
-    /**
-     * @dataProvider serviceExceptionDataProvider
-     *
-     * @param array<mixed> $expectedData
-     */
-    public function testRevokeHandlesException(
-        \Exception|ResponseInterface $httpFixture,
-        int $expectedStatusCode,
-        array $expectedData
-    ): void {
-        $mockingHttpClient = self::getContainer()->get('app.test.mocking_http_client');
-        \assert($mockingHttpClient instanceof HttpClientInterface);
-
-        $httpMockHandler = self::getContainer()->get(MockHandler::class);
-        \assert($httpMockHandler instanceof MockHandler);
-
-        $httpMockHandler->append($httpFixture);
-
-        $token = md5((string) rand());
-        $refreshToken = md5((string) rand());
-
-        self::getContainer()->set(HttpClientInterface::class, $mockingHttpClient);
-
-        $response = $this->applicationClient->makeRevokeRefreshTokenRequest($token, $refreshToken);
-
-        $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function serviceExceptionDataProvider(): array
-    {
-        return array_merge(
-            $this->serviceBadResponseContentTypeDataProvider('user', 'application/json'),
-            $this->serviceHttpFailureDataProvider('user'),
-        );
-    }
-}
